Fix OTJ slow load time
The OTJ page is taking seconds to run through the initial populating of
submission list. Remove the two sequential for loops and gather the
revision information during the loading of the submissions.  The revision
information is then used as the array to sort and the items is sorted
by the same information.